WARNING! Complete overhaul of code!
TagCloudSVG is now a class. This makes is possible to have multiple instances of TagCloudSVG running in parallel.

To create your TagCloudSVG, you should call:

	var myTagCloudSVG = new TagCloudSVG(cloud);

There is a default animate-function, which can be overwritten:

	myTagCloudSVG.animate = function() {
		rotateAndZoom(0.01,0.008,0.005,0.01);
	};

myTagCloudSVG.svg -> The SVG-element
myTagCloudSVG.cloud -> The tagcloud-data
myTagCloudSVG.elements -> Array of the tags' elements
myTagCloudSVG.config -> List of configuration-properties
myTagCloudSVG.drawing -> List of drawing-properties and volatile variables
myTagCloudSVG.addElementToSVG(tag); -> Add an SVG element to the SVG, by referencing an entry in the myTagCloudSVG.cloud array.

Enjoy!